feat: support clipboard text file
