+++ config.toml
@@ -22,6 +22,13 @@ use_local_time_zone = true
 # For example for a collection called `Bing AI images` add the entry "Bing AI images".
 collections_to_include = []
 
+[image_source]
+# The image source to use.
+# Available options are:
+# - api: Uses the collection API to gather image data. Limited to 1000 images.
+# - file: Uses the images_clipboard.txt file to gather image data. Does not contain same thumbnail data as Collection API.
+method = "file"
+
 [debug]
 
 # Enables additional debug statements and debug functionality.
