+++ images_clipboard.txt.example
@@ -1,0 +1,4 @@
+A cat. Image 2 of 4
+https://www.bing.com/images/create/a-cat/1-6591505ae7dc4d86a29b8ecbf2e2ceda?id=ODC6w%2bqsvP%2fqfDLmCeXaTg%3d%3d&view=detailv2&idpp=genimg&idpbck=1&form=BICREC
+www.bing.com
+
